created separate diagnostic application
